Move InterceptingCallInvoker to its own file
